split parsers into separate files, added validation
